// src/app/model/procedure.model.ts
export class ProcedureModel {
    Id: number;
    TransactionNumber: string;
    TransactionDateTime: string;
    TransactionTime: string;
    DICOMFileName: string;
    PatientName: string;
    Gender: string;
    DateOfBirth: string;
    Age: number;
    Particulars: string;
    ModalityId: number;
    Modality: string;
    BodyPartId: number;
    BodyPart: string;
    UserId: number;
    User: string;
    PatientAddress: string;
    ReferringPhysician: string;
    StudyDate: string;
    HospitalNumber: string;
    HospitalWardNumber: string;
    StudyInstanceId: string;
}